requirements.txt
discord.py==1.7.3
git+https://github.com/Rapptz/discord-ext-menus
dnspython==2.1.0
pymongo==3.11.3
requests==2.24.0
Random-Word==1.0.7
PyYAML==5.4.1
pyzt==1.0.1